Refactor element pages into a single layout page
