+++ src/features/Element/MainContent.jsx
@@ -22,7 +22,7 @@ import ExpandableTextBlock from "../../components/ExpandableTextBlock";
 import ContributorCard from "./ContributorCard";
 import { printListWithDelimiter } from "../../helpers/helper";
 import { PeriodAgoText } from "../../utils/PeriodAgoText";
-import { RESOURCE_TYPE_NAMES_PLURAL_FOR_URI } from "../../configs/VarConfigs";
+import { ELEMENT_TYPE_URI_PLURAL } from "../../configs/VarConfigs";
 
 const REACT_FRONTEND_URL = import.meta.env.VITE_REACT_FRONTEND_URL;
 const WEBSITE_TITLE = import.meta.env.VITE_WEBSITE_TITLE;
@@ -44,7 +44,7 @@ export default function MainContent(props) {
   const updateTime = props.updateTime;
   const isLoading = props.isLoading;
 
-  const elementTypePlural = RESOURCE_TYPE_NAMES_PLURAL_FOR_URI[elementType];
+  const elementTypePlural = ELEMENT_TYPE_URI_PLURAL[elementType];
 
   const hasTimestamp = creationTime || updateTime;
   const timePassedText = updateTime
