+++ src/features/ElementGrid/ElementGridLayout.jsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 import { CssVarsProvider as JoyCssVarsProvider } from "@mui/joy/styles";
 import CssBaseline from "@mui/material/CssBaseline";
 
@@ -12,7 +10,7 @@ import Header from "../../components/Layout/Header";
 
 import {
   DEFAULT_BODY_HEIGHT,
-  RESOURCE_TYPE_NAMES_PLURAL_FOR_URI,
+  ELEMENT_TYPE_URI_PLURAL,
 } from "../../configs/VarConfigs";
 
 export default function ElementGridLayout(props) {
@@ -57,7 +55,7 @@ export default function ElementGridLayout(props) {
             }}
           >
             <ElementGrid
-              uriPrefix={"/" + RESOURCE_TYPE_NAMES_PLURAL_FOR_URI[elementType]}
+              uriPrefix={"/" + ELEMENT_TYPE_URI_PLURAL[elementType]}
               elementType={elementType}
               fieldName={fieldName}
               matchValue={matchValue}
